- Make terrain file loader check file size, to improve robustness if   bad file. - Added text entry dialog - Added pilot name, aircraft type and rego to configuration dialog - Added support for team code - Map zoom improvements - Waypoints label in abort fix - Default task at startup fix - Minor UI cleanups (cosmetics) - Logger inactive when in IGC replay mode - Circling wind estimator won't update if less than one fix every 2   seconds. - Zigzag wind estimate inactive when in IGC replay mode - Fixed bug in altair.xci, nearest waypoint details were pan-relative - Analysis dialog: base/ceiling estimation improvements - Task speed now altitude compensated - New task speed instantaneous - All flight statistics retained when exiting XCSoar and loaded at startup,    so previous flight can be reviewed later. - Task is saved when exiting XCSoar and loaded at startup.

+++ Common/Header/leastsqs.h
@@ -55,9 +55,9 @@ class LeastSquares {
 
   double y_ave;
 
-  double xstore[MAX_STATISTICS];
-  double ystore[MAX_STATISTICS];
-  double weightstore[MAX_STATISTICS];
+  float xstore[MAX_STATISTICS];
+  float ystore[MAX_STATISTICS];
+  float weightstore[MAX_STATISTICS];
 
   LeastSquares() {
     Reset();
